langbridge-connectors/langbridge-connector-stripe/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "langbridge-connector-stripe"
version = "0.1.0"
description = "Declarative Stripe SaaS connector example for Langbridge"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["langbridge", "PyYAML>=6.0"]

[project.entry-points."langbridge.connectors"]
stripe = "langbridge_connector_stripe:get_connector_plugin"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
langbridge_connector_stripe = ["manifests/*.yaml"]
